Db. Move and rename Db to SQLIteAdapter
